2.3.1b4
Including changes to mnexec -c to reduce the number of
close() system calls, which may help with performance in
containers on Ubuntu 22.04.

Note that systemd-udevd also seems to be a cause of performance issues
and should probably be disabled while running Mininet.

Adding dbeug logging to test_simpleperf (and net.iperf) because it is
failing intermittently. (Note with packet loss it can fail, but this
is not how we should be testing it.)